requirements.txt
asyncpg
uvicorn
fastapi
python-jose
rich
jsonschema
jinja2
python-multipart
pytest
deepdiff